+++ src/vulncheck.h
@@ -9,33 +9,23 @@ struct TemplatePacket;
 struct MassVulnCheck {
   const char *name;
 
-  /**
-   * A list of default port ranges that should be used in case that none
-   * are specified.
-   */
+  /* A list of default port ranges that should be used in case that none
+   * are specified. */
   const char *ports;
 
-  /**
-   * The hello packet template
-   */
+  /* The hello packet template */
   const unsigned char *packet;
 
-  /**
-   * The hello packet template length
-   */
-  unsigned packet_length;
+  /* The hello packet template lengthn */
+  size_t packet_length;
 
-  /**
-   * Called to change the template based upon the target
-   */
+  /* Called to change the template based upon the target */
   void (*set_target)(struct TemplatePacket *tmpl, const ipv4address_t *ip_them,
                      unsigned port_them, const ipv4address_t *ip_me,
                      unsigned port_me, unsigned seqno, unsigned char *px,
                      size_t sizeof_px, size_t *r_length);
 
-  /**
-   * Called at startup to change the template according to options
-   */
+  /* Called at startup to change the template according to options */
   void (*init)(struct TemplatePacket *tmpl);
 };
